feat: seperate class to delegate becauese of multiple inheritance
